Refactoring and fixes.
- Thieves shadow dance now properly fires for server side and client
  side attacks.
- Client side Cast and Attack and Attack and cast have been condensed
- Fixed issues with Thieves shadow dance for server side attacks.
- Fixed issues with XP gain stopping because of XP being a string not a
  int.

+++ app/Flare/Models/Character.php
@@ -85,7 +85,7 @@ class Character extends Model
         'can_adventure_again_at'      => 'datetime',
         'can_settle_again_at'         => 'datetime',
         'level'                       => 'integer',
-        'xp'                          => 'float',
+        'xp'                          => 'integer',
         'xp_next'                     => 'integer',
         'str'                         => 'integer',
         'dur'                         => 'integer',
